Rename queues file to events
