+++ src/isar/script.py
@@ -12,7 +12,7 @@ from isar.apis.api import API
 from isar.config.keyvault.keyvault_service import Keyvault
 from isar.config.log import setup_loggers
 from isar.config.settings import robot_settings, settings
-from isar.models.communication.queues.queues import Events
+from isar.models.communication.queues.events import Events
 from isar.modules import get_injector
 from isar.robot.robot import Robot
 from isar.services.service_connections.mqtt.mqtt_client import MqttClient
