+++ src/isar/storage/uploader.py
@@ -8,7 +8,7 @@ from typing import List, Union
 from injector import inject
 
 from isar.config.settings import settings
-from isar.models.communication.queues.queues import Events
+from isar.models.communication.queues.events import Events
 from isar.storage.storage_interface import StorageException, StorageInterface
 from robot_interface.models.inspection.inspection import Inspection
 from robot_interface.models.mission.mission import Mission
