+++ tests/isar/services/utilities/test_scheduling_utilities.py
@@ -5,7 +5,7 @@ from fastapi import HTTPException
 from pytest_mock import MockerFixture
 
 from isar.models.communication.queues.queue_timeout_error import QueueTimeoutError
-from isar.models.communication.queues.queues import QueueIO
+from isar.models.communication.queues.events import QueueIO
 from isar.services.utilities.queue_utilities import QueueUtilities
 from isar.services.utilities.scheduling_utilities import SchedulingUtilities
 from isar.state_machine.states_enum import States
